🧹 Refactors media details tables into separate widgets
Moves downloads and formats tables into dedicated stateless widgets
to improve code modularity, readability, and maintainability.
Simplifies the main media details view by delegating table rendering
and logic to new reusable components.
